// src/models/InlineObject20.ts
/**
 * Shifter API
 * Shifter API
 *
 * OpenAPI spec version: latest
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class InlineObject20 {
  'hookEvent'?: string;
  'url'?: string;
  'method'?: string;
  'enabled'?: string;
  'headers'?: any;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'hookEvent',
      baseName: 'hook_event',
      type: 'string',
      format: '',
    },
    {
      name: 'url',
      baseName: 'url',
      type: 'string',
      format: '',
    },
    {
      name: 'method',
      baseName: 'method',
      type: 'string',
      format: '',
    },
    {
      name: 'enabled',
      baseName: 'enabled',
      type: 'string',
      format: '',
    },
    {
      name: 'headers',
      baseName: 'headers',
      type: 'any',
      format: '',
    },
  ];

  static getAttributeTypeMap() {
    return InlineObject20.attributeTypeMap;
  }

  public constructor() {}
}