// src/models/InlineResponse2006.ts
/**
 * Shifter API
 * Shifter API
 *
 * OpenAPI spec version: latest
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class InlineResponse2006 {
  'siteId'?: string;
  'hookEvent'?: string;
  'webhookUrl'?: string;
  'method'?: string;
  'enabled'?: boolean;
  'headers'?: any;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'siteId',
      baseName: 'site_id',
      type: 'string',
      format: '',
    },
    {
      name: 'hookEvent',
      baseName: 'hook_event',
      type: 'string',
      format: '',
    },
    {
      name: 'webhookUrl',
      baseName: 'webhook_url',
      type: 'string',
      format: '',
    },
    {
      name: 'method',
      baseName: 'method',
      type: 'string',
      format: '',
    },
    {
      name: 'enabled',
      baseName: 'enabled',
      type: 'boolean',
      format: '',
    },
    {
      name: 'headers',
      baseName: 'headers',
      type: 'any',
      format: '',
    },
  ];

  static getAttributeTypeMap() {
    return InlineResponse2006.attributeTypeMap;
  }

  public constructor() {}
}