// src/models/InlineResponse2007.ts
/**
 * Shifter API
 * Shifter API
 *
 * OpenAPI spec version: latest
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class InlineResponse2007 {
  'siteId'?: string;
  'hookEvent'?: string;
  'webhookUrl'?: string;
  'result'?: string;
  'jonId'?: string;
  'calledDate'?: string;
  'text'?: string;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'siteId',
      baseName: 'site_id',
      type: 'string',
      format: '',
    },
    {
      name: 'hookEvent',
      baseName: 'hook_event',
      type: 'string',
      format: '',
    },
    {
      name: 'webhookUrl',
      baseName: 'webhook_url',
      type: 'string',
      format: '',
    },
    {
      name: 'result',
      baseName: 'result',
      type: 'string',
      format: '',
    },
    {
      name: 'jonId',
      baseName: 'jon_id',
      type: 'string',
      format: '',
    },
    {
      name: 'calledDate',
      baseName: 'called_date',
      type: 'string',
      format: '',
    },
    {
      name: 'text',
      baseName: 'text',
      type: 'string',
      format: '',
    },
  ];

  static getAttributeTypeMap() {
    return InlineResponse2007.attributeTypeMap;
  }

  public constructor() {}
}